refactor: 마이페이지 목 데이터를 lib/mocks로 이동
