+++ src/lib/mocks/index.ts
@@ -1,1 +1,2 @@
 export * from './chatData';
+export * from './mypageData';
