+++ src/lib/mocks/mypageData.ts
@@ -14,9 +14,6 @@ export type RecordItem = {
   summary: string;
 };
 
-/** 탭 미리보기에서 노출할 개수 (더 보기 클릭 시 전체 목록 페이지로 이동). */
-export const PREVIEW_COUNT = 4;
-
 export const MOCK_BOOKS: RegisteredBook[] = Array.from({ length: 13 }, (_, index) => ({
   id: index + 1,
   title: '셰익스피어의 영혼을 담아낸 평론집',
